test: move openapi tests
