+++ __tests__/cmds/openapi/index.test.ts
@@ -50,6 +50,8 @@ const getCommandOutput = () => {
 const getRandomRegistryId = () => Math.random().toString(36).substring(2);
 
 describe('rdme openapi', () => {
+  let testWorkingDir: string;
+
   beforeAll(() => {
     nock.disableNetConnect();
   });
@@ -57,11 +59,14 @@ describe('rdme openapi', () => {
   beforeEach(() => {
     consoleInfoSpy = vi.spyOn(console, 'info').mockImplementation(() => {});
     consoleWarnSpy = vi.spyOn(console, 'warn').mockImplementation(() => {});
+    testWorkingDir = process.cwd();
   });
 
   afterEach(() => {
     consoleInfoSpy.mockRestore();
     consoleWarnSpy.mockRestore();
+
+    process.chdir(testWorkingDir);
 
     nock.cleanAll();
   });
@@ -428,6 +433,88 @@ describe('rdme openapi', () => {
       return mock.done();
     });
 
+    it('should discover and upload an API definition if none is provided', async () => {
+      const registryUUID = getRandomRegistryId();
+
+      const mock = getAPIMock()
+        .get(`/api/v1/version/${version}`)
+        .basicAuth({ user: key })
+        .reply(200, { version })
+        .post('/api/v1/api-registry', body => {
+          return body.match('form-data; name="spec"');
+        })
+        .reply(201, { registryUUID, spec: { openapi: '3.0.0' } })
+        .get('/api/v1/api-specification')
+        .basicAuth({ user: key })
+        .reply(200, []);
+
+      const postMock = getAPIMockWithVersionHeader(version)
+        .post('/api/v1/api-specification', { registryUUID })
+        .basicAuth({ user: key })
+        .reply(201, { _id: 1 }, { location: exampleRefLocation });
+
+      const spec = 'petstore.json';
+
+      await expect(
+        openapi.run({
+          key,
+          version,
+          workingDirectory: './__tests__/__fixtures__/relative-ref-oas',
+        }),
+      ).resolves.toBe(successfulUpload(spec));
+
+      expect(console.info).toHaveBeenCalledTimes(1);
+
+      const output = getCommandOutput();
+      expect(output).toBe(chalk.yellow(`ℹ️  We found ${spec} and are attempting to upload it.`));
+
+      postMock.done();
+      return mock.done();
+    });
+
+    it('should use specified working directory and upload the expected content', async () => {
+      let requestBody;
+      const registryUUID = getRandomRegistryId();
+
+      const mock = getAPIMock()
+        .get(`/api/v1/version/${version}`)
+        .basicAuth({ user: key })
+        .reply(200, { version })
+        .post('/api/v1/api-registry', body => {
+          requestBody = body.substring(body.indexOf('{'), body.lastIndexOf('}') + 1);
+          requestBody = JSON.parse(requestBody);
+
+          return body.match('form-data; name="spec"');
+        })
+        .reply(201, { registryUUID, spec: { openapi: '3.0.0' } })
+        .get('/api/v1/api-specification')
+        .basicAuth({ user: key })
+        .reply(200, []);
+
+      const postMock = getAPIMockWithVersionHeader(version)
+        .post('/api/v1/api-specification', { registryUUID })
+        .basicAuth({ user: key })
+        .reply(201, { _id: 1 }, { location: exampleRefLocation });
+
+      const spec = 'petstore.json';
+
+      await expect(
+        openapi.run({
+          spec,
+          key,
+          version,
+          workingDirectory: './__tests__/__fixtures__/relative-ref-oas',
+        }),
+      ).resolves.toBe(successfulUpload(spec));
+
+      expect(console.info).toHaveBeenCalledTimes(0);
+
+      expect(requestBody).toMatchSnapshot();
+
+      postMock.done();
+      return mock.done();
+    });
+
     it('should return spec update info for dry run', async () => {
       prompts.inject(['update', 'spec2']);
       const registryUUID = getRandomRegistryId();
@@ -459,6 +546,40 @@ describe('rdme openapi', () => {
       ).resolves.toMatch(`dry run! The API Definition located at ${spec} will update this API Definition ID: spec2`);
 
       mockWithHeader.done();
+      return mock.done();
+    });
+
+    it('should return spec create info for dry run (with working directory)', async () => {
+      const registryUUID = getRandomRegistryId();
+
+      const mock = getAPIMock()
+        .get(`/api/v1/version/${version}`)
+        .basicAuth({ user: key })
+        .reply(200, { version })
+        .post('/api/v1/api-registry', body => {
+          return body.match('form-data; name="spec"');
+        })
+        .reply(201, { registryUUID, spec: { openapi: '3.0.0' } })
+        .get('/api/v1/api-specification')
+        .basicAuth({ user: key })
+        .reply(200, []);
+
+      await expect(
+        openapi.run({
+          key,
+          version,
+          dryRun: true,
+          workingDirectory: './__tests__/__fixtures__/relative-ref-oas',
+        }),
+      ).resolves.toMatch(
+        '🎭 dry run! The API Definition located at petstore.json will be created for this project version: 1.0.0',
+      );
+
+      const output = getCommandOutput();
+      expect(output).toMatch(
+        chalk.yellow('🎭 dry run option detected! No API definitions will be created or updated in ReadMe.'),
+      );
+
       return mock.done();
     });
 
@@ -982,6 +1103,14 @@ describe('rdme openapi', () => {
       mockWithHeader.done();
       return mock.done();
     });
+
+    it('should error if no file was provided or able to be discovered', () => {
+      return expect(openapi.run({ key, version, workingDirectory: 'bin' })).rejects.toStrictEqual(
+        new Error(
+          "We couldn't find an OpenAPI or Swagger definition.\n\nPlease specify the path to your definition with `rdme openapi ./path/to/api/definition`.",
+        ),
+      );
+    });
   });
 
   describe('GHA onboarding E2E tests', () => {
@@ -1235,6 +1364,47 @@ describe('rdme openapi', () => {
       return mock.done();
     });
 
+    it('should create GHA workflow (including workingDirectory)', async () => {
+      const yamlFileName = 'openapi-file-workingdirectory';
+      prompts.inject([true, 'openapi-branch-workingdirectory', yamlFileName]);
+      const registryUUID = getRandomRegistryId();
+
+      const mock = getAPIMock()
+        .get(`/api/v1/version/${version}`)
+        .basicAuth({ user: key })
+        .reply(200, { version })
+        .post('/api/v1/api-registry', body => {
+          return body.match('form-data; name="spec"');
+        })
+        .reply(201, { registryUUID, spec: { openapi: '3.0.0' } })
+        .get('/api/v1/api-specification')
+        .basicAuth({ user: key })
+        .reply(200, []);
+
+      const postMock = getAPIMockWithVersionHeader(version)
+        .post('/api/v1/api-specification', { registryUUID })
+        .basicAuth({ user: key })
+        .reply(201, { _id: 1 }, { location: exampleRefLocation });
+
+      const spec = 'petstore.json';
+
+      await expect(
+        openapi.run({
+          spec,
+          key,
+          version,
+          workingDirectory: './__tests__/__fixtures__/relative-ref-oas',
+        }),
+      ).resolves.toMatchSnapshot();
+
+      expect(yamlOutput).toMatchSnapshot();
+      expect(fs.writeFileSync).toHaveBeenCalledTimes(2);
+      expect(fs.writeFileSync).toHaveBeenNthCalledWith(2, `.github/workflows/${yamlFileName}.yml`, expect.any(String));
+
+      postMock.done();
+      return mock.done();
+    });
+
     it('should reject if user says no to creating GHA workflow', async () => {
       prompts.inject(['create', false]);
       const registryUUID = getRandomRegistryId();
@@ -1360,6 +1530,48 @@ describe('rdme openapi', () => {
 
       putMock.done();
       exampleMock.done();
+      return mock.done();
+    });
+
+    it('should contain request header with correct URL with working directory', async () => {
+      const registryUUID = getRandomRegistryId();
+      const mock = getAPIMock()
+        .get(`/api/v1/version/${version}`)
+        .basicAuth({ user: key })
+        .reply(200, { version })
+        .post('/api/v1/api-registry', body => {
+          return body.match('form-data; name="spec"');
+        })
+        .reply(201, { registryUUID, spec: { openapi: '3.0.0' } })
+        .get('/api/v1/api-specification')
+        .basicAuth({ user: key })
+        .reply(200, []);
+
+      const postMock = getAPIMock({
+        'x-rdme-ci': 'GitHub Actions (test)',
+        'x-readme-source': 'cli-gh',
+        'x-readme-source-url':
+          'https://github.com/octocat/Hello-World/blob/ffac537e6cbbf934b08745a378932722df287a53/__tests__/__fixtures__/relative-ref-oas/petstore.json',
+        'x-readme-version': version,
+      })
+        .post('/api/v1/api-specification', { registryUUID })
+        .basicAuth({ user: key })
+        .reply(201, { _id: 1 }, { location: exampleRefLocation });
+
+      const spec = 'petstore.json';
+
+      await expect(
+        openapi.run({
+          spec,
+          key,
+          version,
+          workingDirectory: './__tests__/__fixtures__/relative-ref-oas',
+        }),
+      ).resolves.toBe(successfulUpload(spec));
+
+      after();
+
+      postMock.done();
       return mock.done();
     });
   });
